move check_auto_installability to support file
